server2/requirements.txt
appdirs==1.4.3
astroid==2.2.5
attrs==19.1.0
black==19.3b0
Click==7.0
ffmpeg-python==0.2.0
Flask==1.1.1
future==0.17.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
pycodestyle==2.5.0
pylint==2.3.1
simplejson==3.16.0
six==1.12.0
toml==0.10.0
typed-ast==1.4.0
Werkzeug==0.15.5
wrapt==1.11.2
youtube-dl==2019.9.1
